# backend/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/quotes": {
      "post": {
        "operationId": "QuotesController_submitRfq",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Submit customer Request For Quotation (RFQ)",
        "tags": ["quotes"]
      },
      "get": {
        "operationId": "QuotesController_listQuotes",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Pagination page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter quotes by customer/dealer user UUID",
            "schema": {
              "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter quotes by status",
            "schema": {
              "example": "PENDING",
              "type": "string",
              "enum": [
                "DRAFT",
                "SUBMITTED",
                "NEGOTIATING",
                "APPROVED",
                "REJECTED",
                "EXPIRED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search keyword matching quoteNumber, customer, or company",
            "schema": {
              "example": "Nam Á",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PaginatedQuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List quotes with filtering and pagination",
        "tags": ["quotes"]
      }
    },
    "/quotes/admin": {
      "post": {
        "operationId": "QuotesController_createAdminQuote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create official B2B quotation (Admin only)",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}": {
      "get": {
        "operationId": "QuotesController_getQuoteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get detailed quote by ID",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}/status": {
      "patch": {
        "operationId": "QuotesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuoteStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update quote status along the state machine (Admin only)",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}/items/{itemId}/price": {
      "put": {
        "operationId": "QuotesController_updateItemPrice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "Quote item UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuoteItemPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update negotiated price for a quote line item (Admin only)",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}/messages": {
      "post": {
        "operationId": "QuotesController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendQuoteMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/QuoteMessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Post a message to quote negotiation timeline",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}/approve-to-order": {
      "post": {
        "operationId": "QuotesController_approveToOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ApproveToOrderResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve quote and convert to order (Admin only)",
        "tags": ["quotes"]
      }
    },
    "/quotes/{id}/export-excel": {
      "get": {
        "operationId": "QuotesController_exportExcel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Excel workbook stream"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download B2B quote Excel (.xlsx) spreadsheet",
        "tags": ["quotes"]
      }
    },
...
  },
  "info": {
    "title": "Hyundai E-Commerce API",
    "description": "Hyundai E-Commerce Platform API for Commercial Vehicles, B2B Dealer Quoting, Warehouse Management, and Multi-channel Payment Processing.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      },
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["success"]
      },
...
      "QuoteMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9d"
          },
          "quoteId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9b"
          },
          "senderId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9c"
          },
          "message": {
            "type": "string",
            "example": "Chào chị, chúng tôi có thể hỗ trợ mức giá 14.500.000 VNĐ."
          },
          "sender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteMessageSenderDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          }
        },
        "required": [
          "id",
          "quoteId",
          "senderId",
          "message",
          "sender",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "QuoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9b"
          },
          "quoteNumber": {
            "type": ["string", "null"],
            "example": "QT-20260904-5892"
          },
          "userId": {
            "type": ["string", "null"],
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9e"
          },
          "customerName": {
            "type": ["string", "null"],
            "example": "Trần Văn Doanh"
          },
          "customerPhone": {
            "type": ["string", "null"],
            "example": "0987654321"
          },
          "customerEmail": {
            "type": ["string", "null"],
            "example": "doanh.tv@gmail.com"
          },
          "companyName": {
            "type": ["string", "null"],
            "example": "Tập đoàn Xây Dựng Số 1"
          },
          "taxId": {
            "type": ["string", "null"],
            "example": "0312345678"
          },
          "shippingAddress": {
            "type": ["string", "null"],
            "example": "Chân công trình Nhà ga T3"
          },
          "status": {
            "type": "string",
            "example": "SUBMITTED",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "NEGOTIATING",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ]
          },
          "subtotalPrice": {
            "type": ["string", "null"],
            "example": "180000000.00"
          },
          "vatRate": {
            "type": ["number", "null"],
            "example": 10
          },
          "vatAmount": {
            "type": ["string", "null"],
            "example": "18000000.00"
          },
          "totalQuotedPrice": {
            "type": ["string", "null"],
            "example": "198000000.00"
          },
          "commercialTerms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteCommercialTermsDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "expirationDate": {
            "format": "date-time",
            "type": ["string", "null"],
            "example": "2026-09-24T08:00:00.000Z"
          },
          "note": {
            "type": ["string", "null"],
            "example": "Ghi chú báo giá"
          },
          "orderId": {
            "type": ["string", "null"],
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9f"
          },
          "createdByAdminId": {
            "type": ["string", "null"],
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9c"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-04T08:00:00.000Z"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItemResponseDto"
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteMessageResponseDto"
            }
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteUserSummaryDto"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "quoteNumber",
          "userId",
          "customerName",
          "customerPhone",
          "customerEmail",
          "companyName",
          "taxId",
          "shippingAddress",
          "status",
          "subtotalPrice",
          "vatRate",
          "vatAmount",
          "totalQuotedPrice",
          "commercialTerms",
          "expirationDate",
          "note",
          "orderId",
          "createdByAdminId",
          "createdAt",
          "updatedAt",
          "items",
          "user"
        ]
      },
      "CreateQuoteItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb8f",
            "description": "Catalog product UUID, or null for bespoke custom item"
          },
          "isCustomItem": {
            "type": "object",
            "example": false,
            "default": false,
            "description": "Whether this is a bespoke line item not in catalog"
          },
          "itemName": {
            "type": "string",
            "example": "Máy phát điện Hyundai DHY6000SE",
            "description": "Item name or description"
          },
          "itemModel": {
            "type": "string",
            "example": "DHY6000SE",
            "description": "Manufacturer model code"
          },
          "itemSpecs": {
            "type": "string",
            "example": "5.0kVA - 230V / 50Hz - Chống ồn",
            "description": "Technical specifications summary"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "minimum": 1,
            "description": "Requested quantity"
          },
          "requestedPrice": {
            "type": "string",
            "example": "25000000.00",
            "description": "Customer target/requested unit price"
          }
        },
        "required": ["itemName", "quantity"]
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "example": "Công ty Cổ phần Xây dựng Nam Á",
            "description": "Customer or company contact name"
          },
          "customerPhone": {
            "type": "string",
            "example": "0901234567",
            "description": "Customer contact phone number"
          },
          "customerEmail": {
            "type": "string",
            "example": "contact@nama.vn",
            "description": "Customer contact email"
          },
          "companyName": {
            "type": "string",
            "example": "Công ty Cổ phần Xây dựng Nam Á",
            "description": "Full registered company name"
          },
          "taxId": {
            "type": "string",
            "example": "0312345678",
            "description": "Corporate enterprise tax ID"
          },
          "shippingAddress": {
            "type": "string",
            "example": "Số 45 Lê Duẩn, Quận 1, TP. Hồ Chí Minh",
            "description": "Project or delivery site destination"
          },
          "note": {
            "type": "string",
            "example": "Yêu cầu giao hàng trước ngày 15/10",
            "description": "Customer special requirements or notes"
          },
          "items": {
            "description": "List of requested quote items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateQuoteItemDto"
            }
          }
        },
        "required": ["customerName", "customerPhone", "items"]
      },
...
      "CreateAdminQuoteDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb90",
            "description": "Dealer or customer UUID if registered account"
          },
          "customerName": {
            "type": "string",
            "example": "Công ty TNHH Kỹ Thuật Điện Quang",
            "description": "Customer or enterprise primary contact"
          },
          "customerPhone": {
            "type": "string",
            "example": "0918123456",
            "description": "Customer contact phone number"
          },
          "customerEmail": {
            "type": "string",
            "example": "sales@dienquang.com.vn",
            "description": "Customer contact email"
          },
          "companyName": {
            "type": "string",
            "example": "Công ty TNHH Kỹ Thuật Điện Quang",
            "description": "Enterprise registered corporate entity"
          },
          "taxId": {
            "type": "string",
            "example": "0309988776",
            "description": "Corporate enterprise tax ID"
          },
          "shippingAddress": {
            "type": "string",
            "example": "Khu Công Nghiệp Sóng Thần 2, Dĩ An, Bình Dương",
            "description": "Project handover destination"
          },
          "vatRate": {
            "type": "object",
            "example": 10,
            "default": 10,
            "description": "VAT percentage rate (e.g. 10 or 8)"
          },
          "commercialTerms": {
            "description": "Structured commercial, warranty, and delivery terms",
            "allOf": [
              {
                "$ref": "#/components/schemas/CommercialTermsDto"
              }
            ]
          },
          "note": {
            "type": "string",
            "example": "Báo giá áp dụng theo chính sách đại lý cấp 1",
            "description": "Internal sales or admin remarks"
          },
          "expirationDate": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-30T00:00:00.000Z",
            "description": "Explicit quote expiration timestamp"
          },
          "items": {
            "description": "Quotation line items with pricing and discounts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminQuoteItemInputDto"
            }
          }
        },
        "required": ["customerName", "customerPhone", "items"]
      },
      "PaginatedQuoteResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 10
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": ["items", "total", "page", "limit"]
      },
      "UpdateQuoteStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "APPROVED",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "NEGOTIATING",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ],
            "description": "Target quotation workflow status"
          }
        },
        "required": ["status"]
      },
      "UpdateQuoteItemPriceDto": {
        "type": "object",
        "properties": {
          "agreedPrice": {
            "type": "string",
            "example": "26500000.00",
            "description": "Agreed renegotiated unit price for quote line item (VND)"
          }
        },
        "required": ["agreedPrice"]
      },
      "SendQuoteMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Chúng tôi đề xuất chiết khấu thêm 2% nếu quý khách lấy số lượng từ 5 máy trở lên.",
            "description": "Negotiation message content"
          }
        },
        "required": ["message"]
      },
      "ApproveToOrderResponseDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9f"
          },
          "quoteId": {
            "type": "string",
            "example": "019fa8bc-8f4d-7000-b366-e691f45cfb9b"
          },
          "status": {
            "type": "string",
            "example": "APPROVED",
            "enum": [
              "DRAFT",
              "SUBMITTED",
              "NEGOTIATING",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ]
          }
        },
        "required": ["orderId", "quoteId", "status"]
      },
...
    }
  }
}